requirements.txt
Flask==2.3.2
requests==2.31.0
pdfminer.six==20221105
